Add call to loop.Run to evaluate all setTimeout/setIntervals methods.
With the addition of the ottoext package which provides a event loop for
usage with otto, we need to call ottoext/loop.Loop.Run method to ensure
all pending tasks are runned within the loop, without this call, all tasks
will be pending and unresolved.

See timer test and loop method:
1. https://github.com/deoxxa/ottoext/blob/master/loop/loop.go#L149
2. https://github.com/deoxxa/ottoext/blob/master/timers/timers_test.go#L17

Related to issue #196.