Refactor JailRuntime and JailCell to expose simpler methods
We need to simplify the interface methods JailRuntime current exposes
as we currently do not make use of it's current exposed methods which
is leading to more interface pollution, though at small scale. We also
are adding mutex coordination which guards all `JailCell.vm` and
`JailCell.loop` calls to ensure all cells are equally guarded for
sequential but concurrent access.

There is an exception to this which is in `Jail.GetCellVM`, which we
will need to decide if this should be made public as is.

Related to issue #196.

Resolves Review comments #2 and #3 from @tiabc.